chore(structure): split document layout footer
