+++ packages/sanity/src/structure/panes/document/document-layout/DocumentLayout.tsx
@@ -1,12 +1,6 @@
-import {
-  DialogProvider,
-  type DialogProviderProps,
-  Flex,
-  PortalProvider,
-  useElementRect,
-} from '@sanity/ui'
+import {DialogProvider, type DialogProviderProps, Flex, useElementRect} from '@sanity/ui'
 import {isHotkey} from 'is-hotkey-esm'
-import {type ReactNode, useCallback, useMemo, useState} from 'react'
+import {useCallback, useMemo, useState} from 'react'
 import {useTranslation} from 'react-i18next'
 import {
   ChangeConnectorRoot,
@@ -15,17 +9,13 @@ import {
   FieldActionsProvider,
   FieldActionsResolver,
   GetFormValueProvider,
-  getSanityCreateLinkMetadata,
   type Path,
   useGlobalCopyPasteElementHandler,
-  useSanityCreateConfig,
   useZIndex,
 } from 'sanity'
 import {styled} from 'styled-components'
 
-import {TooltipDelayGroupProvider} from '../../../../ui-components'
-import {Pane, PaneFooter, usePane, usePaneLayout, usePaneRouter} from '../../../components'
-import {DOCUMENT_PANEL_PORTAL_ELEMENT} from '../../../constants'
+import {Pane, usePaneLayout, usePaneRouter} from '../../../components'
 import {useDocumentIdStack} from '../../../hooks/useDocumentIdStack'
 import {structureLocaleNamespace} from '../../../i18n'
 import {useStructureTool} from '../../../useStructureTool'
@@ -40,9 +30,9 @@ import {DocumentPanel} from '../documentPanel'
 import {DocumentPanelHeader} from '../documentPanel/header'
 import {DocumentActionShortcuts} from '../keyboardShortcuts'
 import {getMenuItems} from '../menuItems'
-import {DocumentStatusBar} from '../statusBar'
 import {useDocumentPane} from '../useDocumentPane'
 import {DocumentLayoutError} from './DocumentLayoutError'
+import {DocumentLayoutFooter} from './DocumentLayoutFooter'
 
 const EMPTY_ARRAY: [] = []
 
@@ -91,9 +81,6 @@ export function DocumentLayout() {
   const {collapsed: layoutCollapsed} = usePaneLayout()
 
   const zOffsets = useZIndex()
-
-  const createLinkMetadata = getSanityCreateLinkMetadata(value)
-  const {startInCreateBanner: StartInCreateBanner} = useSanityCreateConfig().components ?? {}
 
   const [rootElement, setRootElement] = useState<HTMLDivElement | null>(null)
   const [footerElement, setFooterElement] = useState<HTMLDivElement | null>(null)
@@ -178,11 +165,6 @@ export function DocumentLayout() {
     [onPathOpen, onFocus],
   )
 
-  const portalElements = useMemo(
-    () => ({[DOCUMENT_PANEL_PORTAL_ELEMENT]: documentPanelPortalElement}),
-    [documentPanelPortalElement],
-  )
-
   if (!schemaType) {
     return (
       <DocumentLayoutError
@@ -245,32 +227,11 @@ export function DocumentLayout() {
                   rootElement={rootElement}
                   setDocumentPanelPortalElement={setDocumentPanelPortalElement}
                   footer={
-                    // These providers are added because we want the dialogs in `DocumentStatusBar` to be scoped to the document pane
-                    // The portal element comes from `DocumentPanel`.
-                    <PortalProvider __unstable_elements={portalElements}>
-                      <DialogProvider position={DIALOG_PROVIDER_POSITION} zOffset={zOffsets.portal}>
-                        <PaneFooter ref={setFooterElement} padding={1}>
-                          {StartInCreateBanner && (
-                            <ShowWhenPaneOpen>
-                              <StartInCreateBanner
-                                document={value}
-                                documentId={documentId}
-                                documentType={schemaType}
-                                documentReady={ready}
-                                isInitialValueLoading={!!isInitialValueLoading}
-                                panelPortalElementId={DOCUMENT_PANEL_PORTAL_ELEMENT}
-                              />
-                            </ShowWhenPaneOpen>
-                          )}
-                          <TooltipDelayGroupProvider>
-                            <DocumentStatusBar
-                              actionsBoxRef={setActionsBoxElement}
-                              createLinkMetadata={createLinkMetadata}
-                            />
-                          </TooltipDelayGroupProvider>
-                        </PaneFooter>
-                      </DialogProvider>
-                    </PortalProvider>
+                    <DocumentLayoutFooter
+                      documentPanelPortalElement={documentPanelPortalElement}
+                      setFooterElement={setFooterElement}
+                      setActionsBoxElement={setActionsBoxElement}
+                    />
                   }
                 />
               </StyledChangeConnectorRoot>
@@ -281,13 +242,4 @@ export function DocumentLayout() {
       </FieldActionsProvider>
     </GetFormValueProvider>
   )
-}
-
-/**
- * Prevents whatever is inside of it from rendering when the pane is collapsed.
- * Needed locally as DocumentLayout does lives outside PaneContext, but is provided _somewhere_ within it.
- */
-function ShowWhenPaneOpen(props: {children: ReactNode}) {
-  const {collapsed} = usePane()
-  return collapsed ? null : props.children
 }
